# index.js
// Import required modules
const express = require("express");
const cors = require("cors");
const path = require("path");
const fs = require("fs-extra");
require("dotenv").config();

// Import controllers and middleware
const fileController = require("./src/controllers/file-manage-controller");
const authController = require("./src/controllers/auth-controller");
const invitationController = require("./src/controllers/invitation-controller");
const paypalController = require("./src/controllers/paypal-integration");
const stripeController = require("./src/controllers/stripe-controller");
const otaController = require("./src/controllers/ota-controller");
const terminalCredentialController = require("./src/controllers/terminal-credential-controller");
const qpChargeController = require("./src/controllers/qp-charge-controller");
const qpLogController = require("./src/controllers/qp-log-controller");
const { authenticateToken } = require("./src/middleware/auth");
const multer = require("multer");
const swaggerUi = require("swagger-ui-express");
const swaggerDocument = require("./swagger.json");

// Configure multer for file uploads (for dispute evidence)
const storage = multer.diskStorage({
  destination: function (req, file, cb) {
    cb(null, path.join(__dirname, "public", "temp"));
  },
  filename: function (req, file, cb) {
    cb(null, Date.now() + "-" + file.originalname);
  },
});

const uploadEvidence = multer({
  storage: storage,
  limits: {
    fileSize: 10 * 1024 * 1024, // 10MB limit
  },
  fileFilter: function (req, file, cb) {
    // Allow common document types for dispute evidence
    const allowedTypes = [
      "application/pdf",
      "image/jpeg",
      "image/jpg",
      "image/png",
      "image/gif",
      "application/msword",
      "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
      "text/plain",
    ];

    if (allowedTypes.includes(file.mimetype)) {
      cb(null, true);
    } else {
      cb(new Error("Invalid file type for dispute evidence"), false);
    }
  },
});

// Dedicated multer for XLSX/CSV imports (Charge Files & Terminal Credentials)
const uploadChargeFiles = multer({
  storage: storage,
  limits: {
    fileSize: 50 * 1024 * 1024, // 50MB limit for bulk imports
  },
  fileFilter: function (req, file, cb) {
    // Allow XLSX and CSV files for charge imports
    const allowedTypes = [
      "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", // .xlsx
      "application/vnd.ms-excel", // .xls
      "text/csv",
      "text/plain", // Some systems send CSV as plain text
      "application/x-csv",
      "application/csv",
    ];

    if (allowedTypes.includes(file.mimetype)) {
      cb(null, true);
    } else {
      cb(
        new Error("Invalid file type. Please upload XLSX or CSV only."),
        false,
      );
    }
  },
});

// Import database connection
const connectDB = require("./src/config/database");

// Create an instance of Express
const app = express();

// Connect to MongoDB
connectDB();

// Custom HTTP request logging middleware
const requestLogger = (req, res, next) => {
  const startTime = Date.now();
  const timestamp = new Date().toISOString();

  // Store original end method
  const originalEnd = res.end;

  // Override end method to capture response time
  res.end = function (...args) {
    const endTime = Date.now();
    const duration = endTime - startTime;

    // Color coding for status codes
    let statusColor = "";
    if (res.statusCode >= 200 && res.statusCode < 300) {
      statusColor = "\x1b[32m"; // Green for success
    } else if (res.statusCode >= 300 && res.statusCode < 400) {
      statusColor = "\x1b[33m"; // Yellow for redirects
    } else if (res.statusCode >= 400 && res.statusCode < 500) {
      statusColor = "\x1b[31m"; // Red for client errors
    } else if (res.statusCode >= 500) {
      statusColor = "\x1b[35m"; // Magenta for server errors
    }

    // Reset color
    const resetColor = "\x1b[0m";

    // Method color
    const methodColor = "\x1b[36m"; // Cyan for method

    // Call original end method
    originalEnd.apply(this, args);
  };

  next();
};

// Apply request logging middleware
app.use(requestLogger);

// Enable CORS
app.use(cors());

// Stripe webhook endpoint (must be before express.json() middleware)
app.post(
  "/webhook",
  express.raw({ type: "application/json" }),
  stripeController.handleStripeWebhook,
);

// Enable JSON parsing for request bodies
app.use(express.json());

// Serve static files from the 'public' directory
app.use(express.static(path.join(__dirname, "public")));

// Ensure the files directory exists
const filesDir = path.join(__dirname, "public", "files");
fs.ensureDirSync(filesDir);

// Ensure the temp directory exists for dispute evidence uploads
const tempDir = path.join(__dirname, "public", "temp");
fs.ensureDirSync(tempDir);

// Authentication Routes
app.post("/api/auth/register", authController.register);
app.post("/api/auth/login", authController.login);
app.post("/api/auth/verify-otp", authController.verifyOTP);
app.post("/api/auth/resend-otp", authController.resendOTP);
app.post("/api/auth/forgot-password", authController.forgotPassword);
app.post("/api/auth/reset-password/:token", authController.resetPassword);

// Invitation Routes
app.post(
  "/api/invitations/send",
  authenticateToken,
  invitationController.sendInvitation,
);
app.post("/api/invitations/validate", invitationController.validateInvitation);
app.get(
  "/api/invitations/my-invitations",
  authenticateToken,
  invitationController.getMyInvitations,
);
app.post("/api/invitations/complete", invitationController.completeInvitation);

// Protected Routes (require authentication)
app.get("/api/auth/profile", authenticateToken, authController.getProfile);
app.put("/api/auth/profile", authenticateToken, authController.updateProfile);

// File Management API Routes (protected with S3 integration)
app.post(
  "/api/upload",
  authenticateToken,
  fileController.upload.single("file"),
  fileController.uploadFile,
);
app.get("/api/get-row-data", authenticateToken, fileController.getRowData);
app.get(
  "/api/get-stripe-row-data",
  authenticateToken,
  fileController.getStripeRowData,
);
app.get(
  "/api/transaction-history",
  authenticateToken,
  fileController.getTransactionHistory,
);
app.get(
  "/api/get-single-row-data/:documentId",
  authenticateToken,
  fileController.getSingleRowData,
);
app.put(
  "/api/update-sheet-data/:documentId",
  authenticateToken,
  fileController.updateSheet,
);
app.post(
  "/api/create-excel-data",
  authenticateToken,
  fileController.createExcelData,
);
app.get(
  "/api/manual-excel-data",
  authenticateToken,
  fileController.getManualExcelData,
);
app.get("/api/user-files", authenticateToken, fileController.getUserFiles);
app.get("/api/file-headers", authenticateToken, fileController.getFileHeaders);
app.delete(
  "/api/delete-file/:documentId",
  authenticateToken,
  fileController.deleteFile,
);
app.get(
  "/api/files/:uploadId/download",
  fileController.downloadExcelByUploadId,
);

// Upload Management API Routes (protected)
app.get(
  "/api/upload/status/:uploadId",
  authenticateToken,
  fileController.getUploadStatus,
);
app.get(
  "/api/upload/sessions",
  authenticateToken,
  fileController.getUserUploadSessions,
);
app.get(
  "/api/upload/files",
  authenticateToken,
  fileController.getUploadFileSummaries,
);
app.post(
  "/api/upload/resume/:uploadId",
  authenticateToken,
  fileController.resumeUpload,
);
app.delete(
  "/api/upload/cleanup",
  authenticateToken,
  fileController.cleanupFailedUploads,
);
app.delete(
  "/api/upload/delete/:uploadId",
  authenticateToken,
  fileController.deleteUploadById,
);
app.post(
  "/api/upload/archive/:uploadId",
  authenticateToken,
  fileController.archiveFile,
);
app.post(
  "/api/upload/unarchive/:uploadId",
  authenticateToken,
  fileController.unarchiveFile,
);

// PayPal Payment API Routes (protected)
app.post(
  "/api/paypal/process-payment",
  authenticateToken,
  paypalController.processPayment,
);
app.post(
  "/api/paypal/process-bulk-payments",
  authenticateToken,
  paypalController.processBulkPayments,
);
app.get(
  "/api/paypal/payment-details/:documentId",
  authenticateToken,
  paypalController.getPaymentDetails,
);

// PayPal Refund API Routes (protected)
app.post(
  "/api/paypal/process-refund",
  authenticateToken,
  paypalController.processRefund,
);
app.post(
  "/api/paypal/process-bulk-refunds",
  authenticateToken,
  paypalController.processBulkRefunds,
);
app.get(
  "/api/paypal/refund/:refundId",
  authenticateToken,
  paypalController.getRefundDetails,
);

// Stripe Connect API Routes (protected)
app.post(
  "/api/stripe/create-account",
  authenticateToken,
  stripeController.createAccount,
);
app.get(
  "/api/stripe/accounts",
  authenticateToken,
  stripeController.listAccounts,
);
app.get(
  "/api/stripe/account/:accountId",
  authenticateToken,
  stripeController.getAccountById,
);
app.delete(
  "/api/stripe/account/:accountId",
  authenticateToken,
  stripeController.deleteAccount,
);
app.post(
  "/api/stripe/payment",
  authenticateToken,
  stripeController.createSinglePayment,
);

// Stripe Refund API Route (protected)
app.post(
  "/api/stripe/refund",
  authenticateToken,
  stripeController.processStripeRefund,
);

// Stripe Settings API Routes (protected)
app.get(
  "/api/stripe/settings",
  authenticateToken,
  stripeController.getStripeSettings,
);
app.put(
  "/api/stripe/settings",
  authenticateToken,
  stripeController.updateStripeSettings,
);

// Stripe Dispute API Routes (protected)
app.get(
  "/api/stripe/disputes",
  authenticateToken,
  stripeController.listDisputes,
);

app.get(
  "/api/stripe/disputes/stats",
  authenticateToken,
  stripeController.getDisputeStats,
);
app.get(
  "/api/stripe/dispute/:disputeId",
  authenticateToken,
  stripeController.getDisputeDetails,
);
app.post(
  "/api/stripe/upload-evidence",
  authenticateToken,
  uploadEvidence.single("evidence"),
  stripeController.uploadDisputeEvidence,
);
app.post(
  "/api/stripe/submit-evidence",
  authenticateToken,
  uploadEvidence.single("evidence"),
  stripeController.submitDisputeEvidence,
);

// OTA API Routes (protected)
app.get("/api/ota", authenticateToken, otaController.getAllOTAs);
app.get("/api/ota/:id", authenticateToken, otaController.getOTAById);
app.get("/api/ota/name/:name", authenticateToken, otaController.getOTAByName);
app.post("/api/ota", authenticateToken, otaController.createOTA);
app.put("/api/ota/:id", authenticateToken, otaController.updateOTA);
app.delete("/api/ota/:id", authenticateToken, otaController.deleteOTA);
app.patch("/api/ota/:id/restore", authenticateToken, otaController.restoreOTA);
app.post("/api/ota/seed", authenticateToken, otaController.seedOTAData);

// Mount Swagger documentation at /api-docs
app.use(
  "/api-docs",
  swaggerUi.serve,
  swaggerUi.setup(swaggerDocument, {
    customSiteTitle: "QuantumPay APIs",
    swaggerOptions: {
      persistAuthorization: true,
    },
  }),
);

// Admin API Routes (protected)
app.get(
  "/api/admin/excel-data",
  authenticateToken,
  paypalController.getAdminExcelData,
);

// QuantumPay Terminal Credentials Vault (protected)
app.post(
  "/api/terminal-credentials",
  authenticateToken,
  terminalCredentialController.createCredential,
);
app.get(
  "/api/terminal-credentials",
  authenticateToken,
  terminalCredentialController.getCredentials,
);
app.get(
  "/api/terminal-credentials/export",
  authenticateToken,
  terminalCredentialController.exportCredentials,
);
app.get(
  "/api/terminal-credentials/:id",
  authenticateToken,
  terminalCredentialController.getCredentialById,
);
app.patch(
  "/api/terminal-credentials/:id",
  authenticateToken,
  terminalCredentialController.updateCredential,
);
app.delete(
  "/api/terminal-credentials/:id",
  authenticateToken,
  terminalCredentialController.deleteCredential,
);
// Use dedicated uploadChargeFiles multer for credential imports
app.post(
  "/api/terminal-credentials/import",
  authenticateToken,
  uploadChargeFiles.single("file"),
  terminalCredentialController.importCredentials,
);

// QuantumPay Charge Processing (protected)
app.post(
  "/api/qp-charge-files/import",
  authenticateToken,
  uploadChargeFiles.single("file"),
  qpChargeController.importChargeFile,
);
app.get(
  "/api/qp-charge-files",
  authenticateToken,
  qpChargeController.getChargeFiles,
);
app.get(
  "/api/qp-charge-files/:id",
  authenticateToken,
  qpChargeController.getChargeFileById,
);
app.patch(
  "/api/qp-charge-files/:id",
  authenticateToken,
  qpChargeController.updateChargeFile,
);
app.delete(
  "/api/qp-charge-files/:id",
  authenticateToken,
  qpChargeController.deleteChargeFile,
);
app.post(
  "/api/qp-charge-files/:id/process",
  authenticateToken,
  qpChargeController.processChargeFile,
);
app.get(
  "/api/qp-charge-files/:id/export",
  authenticateToken,
  qpChargeController.exportRawInstances,
);
app.get(
  "/api/qp-charge-files/:id/download-compiled",
  authenticateToken,
  qpChargeController.downloadCompiledFile,
);
app.get(
  "/api/qp-charge-files/:id/progress",
  authenticateToken,
  qpChargeController.getChargeFileProgress,
);

// QuantumPay Charge Instances (protected)
app.get(
  "/api/qp-charge-instances",
  authenticateToken,
  qpChargeController.getChargeInstances,
);
app.get(
  "/api/qp-charge-instances/export",
  authenticateToken,
  qpChargeController.exportChargeInstances,
);
app.get(
  "/api/qp-charge-instances/:id",
  authenticateToken,
  qpChargeController.getChargeInstanceById,
);
app.patch(
  "/api/qp-charge-instances/:id",
  authenticateToken,
  qpChargeController.updateChargeInstance,
);
app.delete(
  "/api/qp-charge-instances/:id",
  authenticateToken,
  qpChargeController.deleteChargeInstance,
);
app.post(
  "/api/qp-charge-instances/:id/process",
  authenticateToken,
  qpChargeController.processChargeInstance,
);
// Process a custom list of instances by ID
app.post(
  "/api/qp-charge-instances/process",
  authenticateToken,
  qpChargeController.processMultipleInstances,
);

// QuantumPay Traceability & Logs (protected)
app.get(
  "/api/qp-payment-attempts",
  authenticateToken,
  qpLogController.getPaymentAttempts,
);
app.get(
  "/api/qp-payment-attempts/:id",
  authenticateToken,
  qpLogController.getPaymentAttemptById,
);
app.get("/api/system-logs", authenticateToken, qpLogController.getSystemLogs);

// Health check route
app.get("/api/health", (req, res) => {
  res.status(200).json({
    status: "success",
    message: "Server is running",
    timestamp: new Date().toISOString(),
  });
});

// Error handling middleware
app.use((err, req, res, next) => {
  console.error("❌ Error:", err.stack);
  res.status(500).json({
    status: "error",
    message: "Something went wrong!",
    error:
      process.env.NODE_ENV === "development"
        ? err.message
        : "Internal server error",
  });
});

// 404 handler
app.use("*", (req, res) => {
  res.status(404).json({
    status: "error",
    message: "Route not found",
  });
});

// Define a port
const PORT = process.env.PORT || 3001;

// Start the server
app.listen(PORT, () => {
  console.log(`Server is running on port ${PORT}`);
});